Restructure streamlit app
